Create guess input prompt
Using `cargo run` while adding lines to main.rs helped me understand the expected formatting. Mainly the importance of semicolons, but also that it will fail to compile if there are unused dependancies.
Interesting that variables are immutable by default, also nice to see that intialising with a String is straight forward.
Referencing, with `&` or otherwise, was not something I was expecting and I will need to pay attention to it.